LWMD-16 move search in to page folders and service
