+++ LMWDev_Test/Controllers/SearchController.cs
@@ -3,8 +3,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Moq;
 using Page_Library.Page.Entities.Page.Interface;
+using Page_Library.Page.Entities.SearchResult.Interface;
 using Page_Library.Page.Service.Interface;
-using Page_Library.Search.Entities.SearchResult.Interface;
 using System;
 using System.Collections.Generic;
 using System.Linq;
