+++ Page_Library/Page/Entities/SearchResult/Base/SearchResultBase.cs
@@ -1,10 +1,10 @@
 ﻿using Page_Library.Content.Entities.Content;
 using Page_Library.Content.Entities.Content.Interface;
 using Page_Library.Page.Entities.Page.Interface;
+using Page_Library.Page.Entities.SearchResult.Interface;
 using Page_Library.Search.Entities.SearchResult.DTO;
-using Page_Library.Search.Entities.SearchResult.Interface;
 
-namespace Page_Library.Search.Entities.SearchResult.Base
+namespace Page_Library.Page.Entities.SearchResult.Base
 {
     public abstract class SearchResultBase : ISearchResult
     {
