+++ Page_Library/Page/Entities/SearchResult/Interface/ISearchResult.cs
@@ -1,7 +1,7 @@
 ﻿using Page_Library.Content.Entities.Content;
 using Page_Library.Content.Entities.Content.Interface;
 
-namespace Page_Library.Search.Entities.SearchResult.Interface
+namespace Page_Library.Page.Entities.SearchResult.Interface
 {
     public interface ISearchResult
     {
