+++ Page_Library/Page/Entities/SearchResult/SearchResult.cs
@@ -1,8 +1,8 @@
 ﻿using Page_Library.Page.Entities.Page.Interface;
-using Page_Library.Search.Entities.SearchResult.Base;
+using Page_Library.Page.Entities.SearchResult.Base;
 using Page_Library.Search.Entities.SearchResult.DTO;
 
-namespace Page_Library.Search.Entities.SearchResult
+namespace Page_Library.Page.Entities.SearchResult
 {
     public class SearchResult : SearchResultBase
     {
