+++ Page_Library/Page/Service/Base/PageServiceBase.cs
@@ -1,5 +1,7 @@
 ﻿using Page_Library.Content.Repository.Interface;
 using Page_Library.Page.Entities.Page.Interface;
+using Page_Library.Page.Entities.SearchResult;
+using Page_Library.Page.Entities.SearchResult.Interface;
 using Page_Library.Page.Factory.Interface;
 using Page_Library.Page.Repository.Interface;
 using Page_Library.Page.Service.Interface;
@@ -25,6 +27,34 @@ namespace Page_Library.Page.Service.Base
             results.SetUpPolymorphContentBlocks(_contentRepository, _contentBlockFactory);
             results.Meta.SetContent(_contentRepository);
             return results;
+        }
+
+        private List<ISearchResult> createSearchResults(List<IPage> pages)
+        {
+            var toReturn = new List<ISearchResult>();
+            foreach (IPage page in pages)
+            {
+                page.Meta.SetContent(_contentRepository);
+                var SearchResult = new SearchResult(page);
+                toReturn.Add(SearchResult);
+            }
+
+            return toReturn;
+        }
+
+        public List<ISearchResult> Search(string searchTerm, string category)
+        {
+            try
+            {
+                List<IPage> pages = _pageRepository.GetPages(searchTerm, category);
+                var toReturn = createSearchResults(pages);
+                toReturn.Reverse();
+                return toReturn;
+            }
+            catch (Exception ex)
+            {
+                throw new ApplicationException($"An error occurred while searching with ID: {searchTerm}.", ex);
+            }
         }
     }
 }
